Basic functionality for trading system

+++ Entities/Goods/BaseGood.cs
@@ -1,8 +1,8 @@
 public abstract class BaseGood
 {
     public string Name { get; set; }
-    public decimal PurchasePrice { get; set; }
-    public decimal SellingPrice { get; set; }
+    public int PurchasePrice { get; set; }
+    public int SellingPrice { get; set; }
 
     // This being abstract means other goods will derive from this and might have specific properties or methods.
 }
